requirements.txt
astroid==2.4.1
Bootstrap-Flask==1.3.1
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pylint==2.5.2
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
